mips: bootm: Cast elf entry into unsigned long
Since elf entry type is now a u64 to accomodate both type of elf files
(64 and 32 bits), we need to cast it to the pointer length before
casting it to the pointer type.

+++ arch/mips/lib/bootm.c
@@ -72,7 +72,7 @@ static int do_bootm_elf(struct image_data *data)
 
 	shutdown_barebox();
 
-	entry = (void *)elf->entry;
+	entry = (void *) (unsigned long) elf->entry;
 
 	entry(-2, phys_to_virt((unsigned long)fdt));
